+++ prowler/lib/outputs/finding.py
@@ -315,6 +315,14 @@ class Finding(BaseModel):
                 output_data["resource_line_range"] = check_output.resource_line_range
                 output_data["framework"] = check_output.check_metadata.ServiceName
 
+            elif provider.type == "llm":
+                output_data["auth_method"] = provider.auth_method
+                output_data["account_uid"] = "llm"
+                output_data["account_name"] = "llm"
+                output_data["resource_name"] = check_output.model
+                output_data["resource_uid"] = check_output.model
+                output_data["region"] = check_output.model
+
             # check_output Unique ID
             # TODO: move this to a function
             # TODO: in Azure, GCP and K8s there are findings without resource_name